requirements.txt

streamlit
python-pptx
google-generativeai
watchdog
